use `machinectl` instead of `su` by default
This goes back to the previous commit of fixing TTY pushback
vulnerability. It is probably more secure to use `machinectl` instead of
`su`, and as I said I will go into more details in the readme. This
solution was also suggested by @freedumbo:arcticfoxes.net on matrix. Big
thanks to him.