set absolute paths for all commands to improve security
There might be a possibility for a threat actor to fake the path to
binaries that are run in sedit, so just to make sure an absolute path is
now set for all commands. I have researched the appropriate paths for
most major distributions so there should be no compatibility issues.
This was suggested by chovanecadam (GitHub). Big thanks to him.